improved concurrency, fix corner radius issue

+++ macOS/WritingTools/App/AppDelegate.swift
@@ -373,7 +373,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSWindowDelegate {
     }
 
     private func performRecoveryReset() {
-        let domain = Bundle.main.bundleIdentifier!
+        guard let domain = Bundle.main.bundleIdentifier else { return }
         UserDefaults.standard.removePersistentDomain(forName: domain)
 
         WindowManager.shared.cleanupWindows()
